+++ contracts/VoucherHub.sol
@@ -13,13 +13,14 @@ import {UUPSUpgradeable} from "@openzeppelin/contracts-upgradeable/proxy/utils/U
 
 import {Voucher} from "./beacon/Voucher.sol";
 import {VoucherProxy} from "./beacon/VoucherProxy.sol";
+import {NonTransferableERC20Upgradeable} from "./NonTransferableERC20Upgradeable.sol";
 import {IVoucherHub} from "./IVoucherHub.sol";
 
 contract VoucherHub is
     AccessControlDefaultAdminRulesUpgradeable,
-    ERC20Upgradeable,
     UUPSUpgradeable,
-    IVoucherHub
+    IVoucherHub,
+    NonTransferableERC20Upgradeable
 {
     // Grant/revoke roles through delayed 2 steps process.
     // Used to grant the rest of the roles.
@@ -77,6 +78,7 @@ contract VoucherHub is
         _grantRole(UPGRADE_MANAGER_ROLE, msg.sender);
         _grantRole(ASSET_ELIGIBILITY_MANAGER_ROLE, assetEligibilityManager);
         _grantRole(VOUCHER_MANAGER_ROLE, voucherManager);
+        // This ERC20 is used solely to keep track of the SRLC's accounting in circulation for all emitted vouchers.
         __ERC20_init("iExec Voucher token", "VCHR");
         __UUPSUpgradeable_init();
         VoucherHubStorage storage $ = _getVoucherHubStorage();
@@ -88,30 +90,6 @@ contract VoucherHub is
                 abi.encode($._voucherBeacon) // constructor args
             )
         );
-    }
-
-    /**
-     * @notice VCHR is not transferable.
-     */
-    function transfer(address to, uint256 value) public pure override returns (bool) {
-        to; // Silence unsused
-        value; // variable warnings
-        revert("VoucherHub: Unsupported transfer");
-    }
-
-    /**
-     *
-     * @notice See `transfer` note above.
-     */
-    function transferFrom(
-        address from,
-        address to,
-        uint256 value
-    ) public pure override returns (bool) {
-        from; // Silence
-        to; // unsused variable
-        value; // warning
-        revert("VoucherHub: Unsupported transferFrom");
     }
 
     function createVoucherType(
@@ -225,6 +203,7 @@ contract VoucherHub is
      * address should never be changed.
      * @param owner The address of the voucher owner.
      * @param voucherType The ID of the voucher type.
+     * @param value The amount of SRLC we need to credit to the voucher.
      * @return voucherAddress The address of the created voucher contract.
      */
     function createVoucher(
@@ -241,7 +220,12 @@ contract VoucherHub is
         Voucher(voucherAddress).initialize(owner, address(this), voucherExpiration, voucherType);
         IERC20($._iexecPoco).transfer(voucherAddress, value); // SRLC
         _mint(voucherAddress, value); // VCHR
-        emit VoucherCreated(voucherAddress, owner, voucherExpiration, voucherType);
+        emit VoucherCreated(voucherAddress, owner, voucherExpiration, voucherType, value);
+    }
+
+    function debitVoucher(uint256 debitAmount) external {
+        _burn(msg.sender, debitAmount);
+        emit VoucherDebited(msg.sender, debitAmount);
     }
 
     /**