+++ test/VoucherHub.test.ts
@@ -9,7 +9,7 @@ import { ethers } from 'hardhat';
 import * as commonUtils from '../scripts/common';
 import * as voucherHubUtils from '../scripts/voucherHubUtils';
 import * as voucherUtils from '../scripts/voucherUtils';
-import { IexecPocoMock__factory, Voucher } from '../typechain-types';
+import { IexecPocoMock, IexecPocoMock__factory, Voucher } from '../typechain-types';
 import { VoucherHub } from '../typechain-types/contracts';
 import { random } from './utils/address-utils';
 
@@ -19,9 +19,12 @@ const duration = 3600;
 const voucherValue = 100;
 const asset = random();
 const assetPrice = 1;
+const debitedValue = 25;
+const initVoucherHubBalance = 10 * voucherValue; // arbitrary value, but should support couple voucher creations
 
 describe('VoucherHub', function () {
     let iexecPoco: string;
+    let iexecPocoInstance: IexecPocoMock;
     let voucherHubWithVoucherManagerSigner: VoucherHub;
     let voucherHubWithAssetEligibilityManagerSigner: VoucherHub;
     let voucherHubWithAnyoneSigner: VoucherHub;
@@ -39,7 +42,7 @@ describe('VoucherHub', function () {
             anyone,
         ] = await ethers.getSigners();
         const beacon = await voucherUtils.deployBeaconAndImplementation(owner.address);
-        const iexecPocoInstance = await new IexecPocoMock__factory()
+        iexecPocoInstance = await new IexecPocoMock__factory()
             .connect(owner)
             .deploy()
             .then((x) => x.waitForDeployment());
@@ -54,11 +57,7 @@ describe('VoucherHub', function () {
         voucherHubWithAssetEligibilityManagerSigner = voucherHub.connect(assetEligibilityManager);
         voucherHubWithAnyoneSigner = voucherHub.connect(anyone);
         await iexecPocoInstance
-            .transfer(
-                await voucherHub.getAddress(),
-                10 * // arbitrary value, but should support couple voucher creations
-                    voucherValue,
-            )
+            .transfer(await voucherHub.getAddress(), initVoucherHubBalance)
             .then((tx) => tx.wait());
         return {
             beacon,
@@ -73,7 +72,7 @@ describe('VoucherHub', function () {
     }
 
     describe('Initialize', function () {
-        it('Should initialize', async () => {
+        it('Should initialize', async function () {
             const { beacon, voucherHub, owner, assetEligibilityManager, voucherManager } =
                 await loadFixture(deployFixture);
             const voucherBeaconAddress = await beacon.getAddress();
@@ -109,7 +108,7 @@ describe('VoucherHub', function () {
             expect(expectedHashes).to.include(actualCodeHash);
         });
 
-        it('Should not initialize twice', async () => {
+        it('Should not initialize twice', async function () {
             const { beacon, voucherHub, assetEligibilityManager, voucherManager } =
                 await loadFixture(deployFixture);
 
@@ -125,7 +124,7 @@ describe('VoucherHub', function () {
     });
 
     describe('Upgrade', function () {
-        it('Should upgrade', async () => {
+        it('Should upgrade', async function () {
             const { voucherHub, owner } = await loadFixture(deployFixture);
             const voucherHubAddress = await voucherHub.getAddress();
             const VoucherHubV2Factory = await ethers.getContractFactory('VoucherHubV2Mock', owner);
@@ -139,7 +138,7 @@ describe('VoucherHub', function () {
             expect(await voucherHubV2.foo()).to.equal('bar'); // V2
         });
 
-        it('Should not upgrade when account is unauthorized', async () => {
+        it('Should not upgrade when account is unauthorized', async function () {
             const { voucherHub, anyone } = await loadFixture(deployFixture);
 
             await expect(
@@ -335,27 +334,44 @@ describe('VoucherHub', function () {
         });
     });
 
-    describe('Create voucher', async function () {
-        it('Should create and initialize voucher', async () => {
+    describe('Create voucher', function () {
+        it('Should create and initialize voucher', async function () {
             const { beacon, voucherHub, assetEligibilityManager, voucherManager, voucherOwner1 } =
                 await loadFixture(deployFixture);
             await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
                 description,
                 duration,
             );
+            const expectedVoucherHubInitRlcBalance = initVoucherHubBalance;
             // Create voucher.
+            const voucherHubInitialSrlcBalance = await iexecPocoInstance.balanceOf(
+                voucherHub.getAddress(),
+            );
             const createVoucherTx = await voucherHubWithVoucherManagerSigner
                 .createVoucher(voucherOwner1, voucherType, voucherValue)
                 .then((tx) => tx.wait());
 
+            const voucherHubPostCreationSrlcBalance = await iexecPocoInstance.balanceOf(
+                voucherHub.getAddress(),
+            );
             const voucherAddress = await voucherHub.getVoucher(voucherOwner1);
             const voucher: Voucher = await commonUtils.getVoucher(voucherAddress);
+            const voucherInitialCreditBalance = await voucherHub.balanceOf(voucher.getAddress());
+            const voucherInitialSrlBalance = await iexecPocoInstance.balanceOf(
+                voucher.getAddress(),
+            );
             const voucherAsProxy = await commonUtils.getVoucherAsProxy(voucherAddress);
             const expectedExpiration = await commonUtils.getExpectedExpiration(
                 duration,
                 createVoucherTx,
             );
             // Run assertions.
+            expect(voucherHubInitialSrlcBalance).to.equal(expectedVoucherHubInitRlcBalance);
+            expect(voucherHubPostCreationSrlcBalance).to.equal(
+                voucherHubInitialSrlcBalance - BigInt(voucherValue),
+            );
+            expect(voucherInitialCreditBalance).to.equal(voucherValue);
+            expect(voucherInitialSrlBalance).to.equal(voucherValue);
             // Events.
             await expect(createVoucherTx)
                 .to.emit(voucherAsProxy, 'BeaconUpgraded')
@@ -363,7 +379,13 @@ describe('VoucherHub', function () {
                 .to.emit(voucher, 'OwnershipTransferred')
                 .withArgs(ethers.ZeroAddress, voucherOwner1.address)
                 .to.emit(voucherHub, 'VoucherCreated')
-                .withArgs(voucherAddress, voucherOwner1.address, expectedExpiration, voucherType);
+                .withArgs(
+                    voucherAddress,
+                    voucherOwner1.address,
+                    expectedExpiration,
+                    voucherType,
+                    voucherValue,
+                );
             // Voucher as proxy
             expect(await voucherAsProxy.implementation(), 'Implementation mismatch').to.equal(
                 await beacon.implementation(),
@@ -376,7 +398,7 @@ describe('VoucherHub', function () {
             expect(await voucher.isAccountAuthorized(voucherOwner1.address)).to.be.true;
         });
 
-        it('Should create different vouchers for different accounts with the same config', async () => {
+        it('Should create different vouchers for different accounts with the same config', async function () {
             // Vouchers are created with the same configuration (type, expiration, ...).
             // The goal is to make sure that configuration is not included in the constructor
             // args which would result in different create2 salts.
@@ -392,6 +414,9 @@ describe('VoucherHub', function () {
                 description,
                 duration,
             );
+            const voucherHubInitialSrlBalance = await iexecPocoInstance.balanceOf(
+                voucherHub.getAddress(),
+            );
             // Create voucher1.
             await expect(
                 voucherHubWithVoucherManagerSigner.createVoucher(
@@ -400,6 +425,10 @@ describe('VoucherHub', function () {
                     voucherValue,
                 ),
             ).to.emit(voucherHub, 'VoucherCreated');
+            const voucherHubFirstCreationSrlcBalance = await iexecPocoInstance.balanceOf(
+                voucherHub.getAddress(),
+            );
+
             const voucherAddress1 = await voucherHub.getVoucher(voucherOwner1);
             const voucher1: Voucher = await commonUtils.getVoucher(voucherAddress1);
             // Create voucher2.
@@ -410,8 +439,18 @@ describe('VoucherHub', function () {
                     voucherValue,
                 ),
             ).to.emit(voucherHub, 'VoucherCreated');
+            const voucherHubSecondCreationSrlcBalance = await iexecPocoInstance.balanceOf(
+                voucherHub.getAddress(),
+            );
+
             const voucherAddress2 = await voucherHub.getVoucher(voucherOwner2);
             const voucher2: Voucher = await commonUtils.getVoucher(voucherAddress2);
+            expect(voucherHubFirstCreationSrlcBalance).to.equal(
+                voucherHubInitialSrlBalance - BigInt(voucherValue),
+            );
+            expect(voucherHubSecondCreationSrlcBalance).to.equal(
+                voucherHubFirstCreationSrlcBalance - BigInt(voucherValue),
+            );
 
             expect(voucherAddress1).is.not.equal(voucherAddress2);
             expect(await voucher1.owner()).to.not.equal(await voucher2.owner());
@@ -419,7 +458,7 @@ describe('VoucherHub', function () {
             expect(await voucher1.getVoucherHub()).to.equal(await voucher2.getVoucherHub());
         });
 
-        it('Should create multiple vouchers with the correct config', async () => {
+        it('Should create multiple vouchers with the correct config', async function () {
             const { voucherHub, voucherOwner1, voucherOwner2 } = await loadFixture(deployFixture);
             // Create type0.
             await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
@@ -429,6 +468,7 @@ describe('VoucherHub', function () {
             const voucherType1 = 1;
             const duration1 = 7200;
             const description1 = 'Long Term Duration';
+            const voucherValue1 = 200;
             // Create type1.
             await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
                 description1,
@@ -445,10 +485,9 @@ describe('VoucherHub', function () {
             const voucherAddress1 = await voucherHub.getVoucher(voucherOwner1);
             const voucher1 = await commonUtils.getVoucher(voucherAddress1);
             const voucherAsProxy1 = await commonUtils.getVoucherAsProxy(voucherAddress1);
-
             // Create voucher2.
             const createVoucherTx2 = await voucherHubWithVoucherManagerSigner
-                .createVoucher(voucherOwner2, voucherType1, voucherValue)
+                .createVoucher(voucherOwner2, voucherType1, voucherValue1)
                 .then((tx) => tx.wait());
             const expectedExpirationVoucher2 = await commonUtils.getExpectedExpiration(
                 duration1,
@@ -457,7 +496,9 @@ describe('VoucherHub', function () {
             const voucherAddress2 = await voucherHub.getVoucher(voucherOwner2);
             const voucher2 = await commonUtils.getVoucher(voucherAddress2);
             const voucherAsProxy2 = await commonUtils.getVoucherAsProxy(voucherAddress2);
-
+            const voucherHubFinalSrlcBalance = await iexecPocoInstance.balanceOf(
+                voucherHub.getAddress(),
+            );
             // Events
             expect(createVoucherTx1)
                 .to.emit(voucherHub, 'VoucherCreated')
@@ -466,6 +507,7 @@ describe('VoucherHub', function () {
                     voucherOwner1.address,
                     expectedExpirationVoucher1,
                     voucherType,
+                    voucherValue,
                 );
             expect(createVoucherTx2)
                 .to.emit(voucherHub, 'VoucherCreated')
@@ -474,6 +516,7 @@ describe('VoucherHub', function () {
                     voucherOwner2.address,
                     expectedExpirationVoucher2,
                     voucherType1,
+                    voucherValue1,
                 );
             // Voucher as proxy
             expect(
@@ -496,9 +539,14 @@ describe('VoucherHub', function () {
             );
             expect(await voucher1.getType(), 'Voucher 1 type mismatch').to.equal(voucherType);
             expect(await voucher2.getType(), 'Voucher 2 type mismatch').to.equal(voucherType1);
+            expect(await voucherHub.balanceOf(voucher1.getAddress())).to.equal(voucherValue);
+            expect(await voucherHub.balanceOf(voucher2.getAddress())).to.equal(voucherValue1);
+            expect(voucherHubFinalSrlcBalance).to.equal(
+                initVoucherHubBalance - voucherValue - voucherValue1,
+            );
         });
 
-        it('Should not create more than 1 voucher for the same account', async () => {
+        it('Should not create more than 1 voucher for the same account', async function () {
             const { voucherHub, assetEligibilityManager, voucherManager, voucherOwner1 } =
                 await loadFixture(deployFixture);
             await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
@@ -523,7 +571,7 @@ describe('VoucherHub', function () {
             ).to.be.revertedWithoutReason();
         });
 
-        it('Should not create more than 1 voucher for the same account with different config', async () => {
+        it('Should not create more than 1 voucher for the same account with different config', async function () {
             const { voucherHub, assetEligibilityManager, voucherManager, voucherOwner1 } =
                 await loadFixture(deployFixture);
             await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
@@ -555,11 +603,11 @@ describe('VoucherHub', function () {
             ).to.be.revertedWithoutReason();
         });
 
-        it('Should not create voucher when initialization fails', async () => {
+        it('Should not create voucher when initialization fails', async function () {
             // TODO
         });
 
-        it('Should not initialize voucher more than once', async () => {
+        it('Should not initialize voucher more than once', async function () {
             const { voucherHub, assetEligibilityManager, voucherManager, voucherOwner1 } =
                 await loadFixture(deployFixture);
             await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(
@@ -588,7 +636,7 @@ describe('VoucherHub', function () {
             ).to.be.revertedWithCustomError(voucher, 'InvalidInitialization');
         });
 
-        it('Should not create voucher when not authorized', async () => {
+        it('Should not create voucher when not authorized', async function () {
             const { voucherHub, voucherOwner1, anyone } = await loadFixture(deployFixture);
             // Create voucher.
             await expect(
@@ -596,7 +644,7 @@ describe('VoucherHub', function () {
             ).to.be.revertedWithCustomError(voucherHub, 'AccessControlUnauthorizedAccount');
         });
 
-        it('Should not create voucher when voucher type ID is out of bounds', async () => {
+        it('Should not create voucher when voucher type ID is out of bounds', async function () {
             const { voucherHub, voucherManager, voucherOwner1 } = await loadFixture(deployFixture);
             const outOfBoundsTypeID = 999;
             // Create voucher.
@@ -729,6 +777,38 @@ describe('VoucherHub', function () {
         });
     });
 
+    describe('Debit voucher', function () {
+        let [voucherOwner1, voucherOwner2, voucher, anyone]: SignerWithAddress[] = [];
+        let voucherHub: VoucherHub;
+
+        beforeEach(async function () {
+            ({ voucherHub, voucherOwner1, voucherOwner2, anyone } =
+                await loadFixture(deployFixture));
+            // Create voucher type
+            await voucherHubWithAssetEligibilityManagerSigner
+                .createVoucherType(description, duration)
+                .then((tx) => tx.wait());
+            // Create voucher
+            voucher = await voucherHubWithVoucherManagerSigner
+                .createVoucher(voucherOwner1, voucherType, voucherValue)
+                .then((tx) => tx.wait())
+                .then(() => voucherHub.getVoucher(voucherOwner1))
+                .then((voucherAddress) => ethers.getImpersonatedSigner(voucherAddress));
+        });
+
+        it('Should debit voucher', async function () {
+            const voucherInitialCreditBalance = await voucherHub.balanceOf(voucher.address);
+            await expect(await voucherHub.connect(voucher).debitVoucher(debitedValue))
+                .to.emit(voucherHub, 'Transfer')
+                .withArgs(voucher.address, ethers.ZeroAddress, debitedValue)
+                .to.emit(voucherHub, 'VoucherDebited')
+                .withArgs(voucher.address, debitedValue);
+            expect(await voucherHub.balanceOf(voucher.address)).equals(
+                voucherInitialCreditBalance - BigInt(debitedValue),
+            );
+        });
+    });
+
     describe('Get voucher', function () {
         it('Should return address 0 when voucher is not created', async function () {
             const { voucherHub, owner } = await loadFixture(deployFixture);
@@ -736,18 +816,25 @@ describe('VoucherHub', function () {
         });
     });
 
-    describe('ERC20', function () {
+    describe('NonTransferableERC20Upgradeable', function () {
         it('Should not transfer', async function () {
             const { voucherHub, anyone } = await loadFixture(deployFixture);
             await expect(voucherHub.transfer(anyone, 0)).to.be.revertedWith(
-                'VoucherHub: Unsupported transfer',
+                'NonTransferableERC20Upgradeable: Unsupported transfer',
+            );
+        });
+
+        it('Should not approve', async function () {
+            const { voucherHub, anyone } = await loadFixture(deployFixture);
+            await expect(voucherHub.approve(anyone, 0)).to.be.revertedWith(
+                'NonTransferableERC20Upgradeable: Unsupported approve',
             );
         });
 
         it('Should not transferFrom', async function () {
             const { voucherHub, anyone } = await loadFixture(deployFixture);
             await expect(voucherHub.transferFrom(anyone, anyone, 0)).to.be.revertedWith(
-                'VoucherHub: Unsupported transferFrom',
+                'NonTransferableERC20Upgradeable: Unsupported transferFrom',
             );
         });
     });